implementantod teste unitarios

+++ src/main/java/origami_flow/salgado_trancas_api/service/MetricaService.java
@@ -45,11 +45,11 @@ public class MetricaService {
         return totalVendas;
     }
 
-    private Integer calcularTaxaDeAgendamentoNoMes(int mes, int ano) {
-        Integer novosClientes = clienteService.clientesNovosNoMes(mes, ano);
-        Integer agendamentosDeNovosCliente = atendimentoRealizadoService.buscarNumeroDeAtendimentoRealizadoComClientesNovos(mes, ano);
-        if (novosClientes == 0) return 0;
-        return (agendamentosDeNovosCliente/novosClientes) * 100;
+    private Double calcularTaxaDeAgendamentoNoMes(int mes, int ano) {
+        Double novosClientes = (double) clienteService.clientesNovosNoMes(mes, ano);
+        Double agendamentosDeNovosCliente = (double) atendimentoRealizadoService.buscarNumeroDeAtendimentoRealizadoComClientesNovos(mes, ano);
+        if (novosClientes == 0) return 0.0;
+        return (double) (agendamentosDeNovosCliente/novosClientes) * 100;
     }
 
     private Double calcularLucroDoMes(int mes, int ano) {
